Attempt to write test for input field
I gave up on enzyme, I don't think it was the right tool for the
job.

This led me to give this react-testing-library a try.

I like that it is letting me test a change in the input changes
and then triggering a submission to see if the baseui to final-form
integration is working.